remove types that don't need to be global
